Save real time plot after each time step as images and make movie from it at the end
Previously, the real time plots movie was saved in a separate process. This caused inconsistent framerates as the separate process captured the system state after fixed periods. Additionally, the movie could not be made in the jupyter notebook as a new process cannot be launched on windows from jupyter.

With the current modification, all the plots are saved manually after every time step in the "images" folder. At the end of simulation opencv is used to create a movie from the images. This provides a consistent framerates as well as possibility to make movie in jupyter. However, this may need more disc space for saving lots of images