introduce separate parameter for saving of frames in the same process
Previously the saving of plots movie was done in the same process as the main simulator. The saving of frame after each time step slows the simulation.

Therefore, we keep the older implementation of recording the movie in the separate process using matplotlib animate method. Now distinction is made between "save_live_plot" and "save_frames". The earlier saves makes movie of the real time plots in a separate process, which can lead to inconsistent update rate depending of simulated scenarios.