Use the abstract class Visitable instead of ServiceStation for time-invariant visitable locations
Previously the ServiceStation class was directly used in the SimulatorClass. This made the time-invariant geographical locations (that remain same during simulations) limited to ServiceStations only.

The new implementation uses the abstraction Visitable for time-invariant StationaryObjects. The SimulatorClass gets these objects through "generate_visitable-objects" method that can return any subclass of Visitable.
Thus, now the SImulatorClass is independent of the exact definition of the Visitable, which the user can define according to his requirements.
The LivePlot class is also modified for general visitable objects.