BUGFIX: remove uptil a visitable object is found for generating optimization copy of the Car
The Car object was ignoring the time-invariant visitable object in its path for generating optimization copy. This was causing the time-invariant visitable object being visited multiple times.

Now the optimization (moving backward from the car path) removes everything till one of the following is fulfilled
1) The Servable object has been optimized the maximum number of times
2) A time-invariant visitable object is found
If any of the above condition is found, then that point is marked as the car origin for optimization copy.