Add arguement to OSMNxGraphRouter to not include time and distance matrices inside pickle
The computed time and distance matrices are usually large in size. Including them inside pickle increases the size of pickled router file significantly.

We add an attribute "full_pickle" to the class which is if False does not include the time and distance matrices inside pickle. The time and distance matrices are manually loaded at the time of unpickling.
If they are available in the "save_folder", then they are loaded from there instead of recalculation