Add custom getstate and setstate to data reader
The data reader has Settings object as attribute which cannot be pickled because the superclass of Settings is Dynaconf (a module object). Thus, data reader cannot be pickled using default implementation.

The customer getstate saves the paths to settingsfiles and settings as dictionary. In setstate a new Settings object is created from the save settings files paths and dictionary