add arguement "resume_state" for resuming simulation from last state and separate generators initialization into a method
Since the generators have to be reinitialized if the state is resumed or unpickled from save state. So for ease we separate their initialization into separate method "initialize_generator_functions".
Add arguement "resume_state" which starts the simulation from the last state. This is to be used after unpickling already saved backup state.