DEBUG: the optimization copy of car should include boarding and disembark time
The availability time "car.treq" was previoudly directly set to the "reach time" of the node. This would allow the amount of time a car stay at the node included inside the treq.
Therefore, the treq is calculated using the estimate_leave_time method of the stationary node. This automatically integrate the boarding and disembark times in the car.treq.