DEBUG: use geographical_point as key for visited locations instead of string representation of visited node to avoid duplicate marking of visited geographical point
With the introduction of disembarking and boarding times, the car might have to wait at a geographical point across time steps. The keys in the simulation_delay dictionary (which uses string representatio of nodes) are used to avoid the duplicate marking of the visited nodes. However, due to the scaling of the time matrix the reach times of the scheduled nodes changes, leading to changes in the string representation of nodes each time time matrix is scaled. This was raising "key not found" error due to duplicate marking of nodes (especially for the drop off points).

To solve the above issue, we directly use the string representation of the geographical point as key in the simulation_delay dict. This key does not change with scaling of the time matrix.