Use already present results folder is the last simulation was not completed
Previously, the if the results folder already existed, a numerical number was added to the end of the results folder. This caused creation of a new folder each time a simulation was restarted.
Now the simulator checks if the simulator.csv exists in the results folder. If it exists, then it appends numerical number at the end otherwise it uses the already existing folder.